removed base URL from ApiService, configured it in Ktor client for safety
The base URL for API requests has been removed from the `ApiService` class.
The Ktor client is now configured to use a base URL in `DefaultRequest`, that is provided by the `NativeLib` and loaded from the native library.
This allows to dynamically set the base URL.
Additionally, the logging interceptor was added to the ktor client.
Also OkHttp dependencies has been moved to the common dependencies.